Linux_tht4/requirements.txt
requests
mysql-connector-python
python-dotenv
plotly
